fixes #175
