+++ apstools/utils.py
@@ -49,6 +49,7 @@ import pyRestTable
 import re
 import smtplib
 import subprocess
+import threading
 import time
 import xlrd
 import zipfile
@@ -225,6 +226,29 @@ def pairwise(iterable):
     """
     a = iter(iterable)
     return zip(a, a)
+
+
+def run_in_thread(func):
+    """
+    (decorator) run ``func`` in thread
+    
+    USAGE::
+
+       @run_in_thread
+       def progress_reporting():
+           logger.debug("progress_reporting is starting")
+           # ...
+       
+       #...
+       progress_reporting()   # runs in separate thread
+       #...
+
+    """
+    def wrapper(*args, **kwargs):
+        thread = threading.Thread(target=func, args=args, kwargs=kwargs)
+        thread.start()
+        return thread
+    return wrapper
 
 
 def split_quoted_line(line):
